## Why is this change being made?
Adding power CLI framework implementation, using interfaces exposed via dfwk.

This PR implements the design pattern for the power CLI using async dfwk queues to request data from the power service.
In addition, most of the `pwr cfg` CLI commands are implemented.

## What is changed?
The following updates have been made:
1. A `power_runconfig_get_element` API is added to power_runconfig to fetch chunks of data from the power runconfig data structure to pass on to external interfaces. The supported list of data chunks is defined in a dictionary `power_runconfig_dictionary`.
2. A Power CLI module is created that interacts with the power service via DriverFramework. It supports both synchronous calls (via `power_service_dispatch_sync`) as well as an asynchronous interface via the default message queue on the power_svc device. This queue over the default interface is shared with the SSI service, both interacting with the power service.
3. RequestTypes specific to the CLI interface are created (these do not overlap with the RequestTypes defined for the SSI service), and these new RequestTypes are handled in the `power_service_dispatch_async` handler. Upon receiving a CLI request over the queue, the power service returns a pointer to the requested data and completes the async request.
4. Completion of the async request triggers the callback registered with the DFWK Queue when the CLI command was called. This callback prints the requested data, using the pointer passed back from the power service.
5. Requesting any data that is not registered with the `power_runconfig_get_element` API causes an `Unsupported sub command` response.

## How is the change being tested?

- [ ] No test, document update only
- [X] Unit tests pass
- [ ] Integration tests pass
- [x] Manual tests pass (add test specifics): Manual command submission on SVP
- [ ] Others, explain:

### Test Platform

- [x] SVP
- [ ] FPGA (Add FPGA Config)
- [ ] Silicon
- [ ] Others, explain:

## Notes

## Risks

Related work items: #1491013

+++ src/libs/corebits/inc/corebits.h
@@ -22,7 +22,7 @@
 
 // helpers for printf
 #define COREBITS_FMT_STR            "%x:%08x:%08x"
-#define COREBITS_FMT_DATA(corebits) corebits.bits[2], corebits.bits[1], corebits.bits[0]
+#define COREBITS_FMT_DATA(corebits) (unsigned int)corebits.bits[2], (unsigned int)corebits.bits[1], (unsigned int)corebits.bits[0]
 
 // helpers for init
 #define COREBITS_INIT_UINT32(core0_31, core32_63, core64_95) \
